requirements.txt
tensorflow=2.8.0
keras=2.8.0
numpy=1.22.3
matplotlib=3.5.1
fiftyone=0.17.2
opencv-python=4.5.5.64
pillow=9.1.0
beautifulsoup4=4.11.1